+++ src/object/every.js
@@ -1,12 +1,15 @@
-define(['./forOwn'], function(forOwn) {
+define(['./forOwn', '../function/makeIterator_'], function(forOwn, makeIterator) {
 
     /**
      * Object every
      */
     function every(obj, callback, thisObj) {
+        callback = makeIterator(callback);
         var result = true;
         forOwn(obj, function(val, key) {
-            if (callback.call(thisObj, val, key, obj) === false) {
+            // we consider any falsy values as "false" on purpose so shorthand
+            // syntax can be used to check property existence
+            if (!callback.call(thisObj, val, key, obj)) {
                 result = false;
                 return false; // break
             }