+++ tests/spec/collection/spec-filter.js
@@ -34,6 +34,42 @@ define(['mout/collection/filter'], function(filter){
             expect( result ).toEqual( ['a', 'c'] );
         });
 
-    });
 
+        it('should support shorthand syntax', function () {
+            var obj = {
+                '0' : {foo:'bar', lorem:'ipsum', id:1},
+                '1' : {foo:'bar', lorem:'ipsum', id:2},
+                '2' : {foo:'bar', lorem:'ipsum', id:4}
+            };
+            var arr = [obj[0], obj[1], obj[2]];
+
+            expect( filter(obj, {foo:'bar', lorem:'ipsum'}) ).toEqual( [obj[0], obj[1], obj[2]] );
+            expect( filter(obj, {lorem:'ipsum', id:1}) ).toEqual( [obj[0]] );
+            expect( filter(obj, {amet:123}) ).toEqual( [] );
+
+            expect( filter(arr, {foo:'bar', lorem:'ipsum'}) ).toEqual( [obj[0], obj[1], obj[2]] );
+            expect( filter(arr, {lorem:'ipsum', id:1}) ).toEqual( [obj[0]] );
+            expect( filter(arr, {amet:123}) ).toEqual( [] );
+        });
+
+
+        it('should allow string shorthand syntax', function () {
+            var obj = {
+                '0' : {foo:'bar', lorem:'ipsum', id:1},
+                '1' : {foo:'bar', lorem:'ipsum', id:2},
+                '2' : {foo:'bar', lorem:'ipsum', id:0}
+            };
+            var arr = [obj[0], obj[1], obj[2]];
+
+            expect( filter(obj, 'foo') ).toEqual( [obj[0], obj[1], obj[2]] );
+            expect( filter(obj, 'id') ).toEqual( [obj[0], obj[1]] );
+            expect( filter(obj, 'amet') ).toEqual( [] );
+
+            expect( filter(arr, 'foo') ).toEqual( [obj[0], obj[1], obj[2]] );
+            expect( filter(arr, 'id') ).toEqual( [obj[0], obj[1]] );
+            expect( filter(arr, 'amet') ).toEqual( [] );
+        });
+
+
+    });
 });